% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_SNEE.R
\name{scale_colour_SNEE}
\alias{scale_colour_SNEE}
\title{SNEE colour scale}
\usage{
scale_colour_SNEE(
  palette = c("greens", "purples", "pinks", "blues", "yellows", "diverge", "qual"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Name of palette.}

\item{reverse}{Sets order of colours. To reverse default order use \code{reverse = TRUE}.}

\item{...}{Additional arguments passed to the ggplot function.}
}
\value{
a scale object to be used with a ggplot object
}
\description{
Generates a colour scale for ggplot using SNEE colour palettes
}
\examples{
\dontrun{# Add SNEE colour scale to a plot
library(ggplot2)
library(SNEEstyle)
ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, colour = Species)) +
geom_point() +
scale_colour_SNEE("greens") +
SNEE_theme()
# or, reverse the scale
library(ggplot2)
library(SNEEstyle)
ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length, colour = Species)) +
geom_point() +
scale_colour_SNEE("greens", reverse = TRUE) +
SNEE_theme()}
}
