% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_SNEE.R
\name{scale_fill_SNEE}
\alias{scale_fill_SNEE}
\title{SNEE fill scale}
\usage{
scale_fill_SNEE(
  palette = c("greens", "purples", "pinks", "blues", "yellows", "diverge", "qual"),
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Name of palette.}

\item{reverse}{Sets order of colours. To reverse default order use \code{reverse = TRUE}.}

\item{...}{Additional arguments passed to the ggplot function.}
}
\value{
a scale object to be used with a ggplot object
}
\description{
Generates a fill colour scale for ggplot using SNEE colour palettes
}
\examples{
\dontrun{
# Add SNEE fill scale to a plot
library(ggplot2)
library(SNEEstyle)
ggplot(iris, aes(x = Sepal.Width, fill = Species)) +
geom_bar(stat = "count", position = "dodge") +
scale_fill_SNEE("greens") +
SNEE_theme()
# or, reverse the scale
library(ggplot2)
library(SNEEstyle)
ggplot(iris, aes(x = Sepal.Width, fill = Species)) +
geom_bar(stat = "count", position = "dodge") +
scale_fill_SNEE("greens", reverse = TRUE) +
SNEE_theme()}
}
