% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNEE_colour.R
\name{SNEE_colour}
\alias{SNEE_colour}
\title{SNEE colour}
\usage{
SNEE_colour(
  ...,
  named = FALSE,
  palette = c(NA, "greens", "purples", "pinks", "blues", "yellows", "diverge", "qual")
)
}
\arguments{
\item{...}{individual colours that you wish to get the values of}

\item{named}{specify whether colour names should be included (default = FALSE)}

\item{palette}{a name of a palette to select the colours of}
}
\value{
a named vector of RGB colours in hexadecimal form
}
\description{
This function returns the colours that are used in the SNEE theme as
a named vector of RGB values in hexadecimal form.

If no arguments are passed to the function, then all of the colours are
returned. If only specific colours are required, then the names of the
colours that are required can be passed as strings to the function, and only
those colours will be returned.

If a palette is selected then no arguments should be passed to ...

Only one palette can be selected at a time.
}
\examples{
# show all of the colours
SNEE_colour()
# or, just show some colours
SNEE_colour("green")
# or, select a single palette
SNEE_colour(palette = "greens")
}
