% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNEE_logo.R, R/SNEE_quarto_setup.R
\name{SNEE_logo}
\alias{SNEE_logo}
\alias{SNEE_quarto_setup}
\title{SNEE logo}
\usage{
SNEE_logo(size = "300")

SNEE_quarto_setup(
  type = c("document", "presentation"),
  format = c("html", "office"),
  file_path,
  file_name
)
}
\arguments{
\item{size}{The size of the logo to be printed in pixels, default = 300px.}
}
\value{
A magick-image object
}
\description{
Pulls SNEE logo from package files for use as an image
}
\examples{
\dontrun{# Display SNEE logo in image viewer
library(SNEEstyle)
SNEE_logo()}
\dontrun{# Create quarto html presentation "draft1.qmd" in "outputs" folder of project
library(SNEEstyle)
SNEE_quarto_setup(type = "presentation", format = "html", "outputs", "draft1")}
}
