% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNEE_quarto_setup.R
\name{SNEE_quarto_setup}
\alias{SNEE_quarto_setup}
\title{SNEE quarto setup}
\arguments{
\item{type}{The quarto type, either document or presentation.}

\item{format}{The format of the final presentation, either html or office (ppt/doc).}

\item{file_path}{The folder where the template and associated files should be stored. This can be an existing or new folder.}

\item{file_name}{The name of the quarto template created.}
}
\value{
Quarto template ready for editing.
}
\description{
Creates a quarto document or presentation template and stores in new or existing project folder
}
