% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNEE_theme.R
\name{SNEE_theme}
\alias{SNEE_theme}
\title{Creates the SNEE Theme for ggplot}
\usage{
SNEE_theme()
}
\arguments{
\item{base_family}{The font to use in this theme}
}
\description{
Creates a base theme for ggplot objects. By default this uses a font that
requires the \link[extrafont]{extrafont} package to be installed.
}
\examples{
\dontrun{
library(ggplot2)
library(SNEEstyle)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point() +
  scale_colour_SNEE("greens") +
  SNEE_theme()
}
}
